Add a general controller program for the raspberry pi
The controller script manages the LEDs, the haptic motor and the audio
channel for Missing Link statues. It does this by sending MQTT
msgs and WLED cmds based off of MQTT msgs received from the Teensy.

It also provides a debug server that allows operators to change certain
values in realtime, like the current operating mode.

Right now the script has not been tested and contains placeholder values
for most of the parameters. The MQTT msg protocol with the Teensy has
not been fully defined. Also the debug server endpoints are not fully
fleshed out. This PR is mainly to demonstrate the overall structure of
the program and its capabilities.