update controller script, touch sensing on teensy
Simplify the main controller script that runs on the Pi. Build in the
latest architecture, ie the Teensy does touch sensing, the Pi plays
audio and the QuinLed board drives the lights and the haptic motors.

Also delete some of the old audio files to save disk space and memory.